refactor: remove unused test module
