--- rust/signed_doc/tests/decoding.rs
@@ -1,1 +1,0 @@
-//! Decoding tests.
