--- rust/signed_doc/tests/signature.rs
@@ -1,12 +1,0 @@
-//! Signature validation tests.
-
-// use catalyst_signed_doc::*;
-
-// #[test]
-// fn test_signature_verification() {
-//     let doc = CatalystSignedDocument::new(); // Assuming new() creates a signed doc
-//     assert!(
-//         doc.verify_signature().is_ok(),
-//         "Signature verification failed"
-//     );
-// }
